Refactored coupled code
