+++ LudoCompose/app/src/main/java/hu/bme/aut/android/ludocompose/ui/model/GameListItemUi.kt
@@ -16,9 +16,7 @@
 
 package hu.bme.aut.android.ludocompose.ui.model
 
-import hu.bme.aut.android.ludocompose.domain.model.GameListItem
 import kotlinx.datetime.LocalDate
-import kotlinx.datetime.toLocalDate
 import java.time.LocalDateTime
 
 data class GameListItemUi internal constructor(
@@ -30,11 +28,4 @@ data class GameListItemUi internal constructor(
         LocalDateTime.now().dayOfMonth
     ).toString(),
     val playerNames: List<String> = emptyList(),
-)
-
-fun GameListItem.toUiModel() = GameListItemUi(
-    id = id,
-    name = name,
-    date = date.toString(),
-    playerNames = playerNames,
 )
